requirements.txt
Flask
Flask-PyMongo
faker
PyJWT
bcrypt
pytest
Flask-CORS